+++ pyoxidizer/src/python_distributions.rs
@@ -67,8 +67,8 @@ pub static PYTHON_DISTRIBUTIONS: Lazy<PythonDistributionCollection> = Lazy::new(
 		PythonDistributionRecord {
 			python_major_minor_version: "3.8".to_string(),
 			location: PythonDistributionLocation::Url {
-				url: "https://github.com/indygreg/python-build-standalone/releases/download/20210103/cpython-3.8.7-x86_64-unknown-linux-gnu-pgo-20210103T1125.tar.zst".to_string(),
-				sha256: "d578b5583cb907eac2a02707e6f43ee0a38b0491f9db5b8e0f23b4f2d3042260".to_string(),
+				url: "https://github.com/indygreg/python-build-standalone/releases/download/20210228/cpython-3.8.8-x86_64-unknown-linux-gnu-pgo-20210228T1503.tar.zst".to_string(),
+				sha256: "b31db3a56a1fb203fc77a013ca0c37d68a90bd03da96eb645ceed3cc176ab230".to_string(),
 			},
 			target_triple: "x86_64-unknown-linux-gnu".to_string(),
 			supports_prebuilt_extension_modules: true,
@@ -76,8 +76,8 @@ pub static PYTHON_DISTRIBUTIONS: Lazy<PythonDistributionCollection> = Lazy::new(
 		PythonDistributionRecord {
 			python_major_minor_version: "3.9".to_string(),
 			location: PythonDistributionLocation::Url {
-				url: "https://github.com/indygreg/python-build-standalone/releases/download/20210103/cpython-3.9.1-x86_64-unknown-linux-gnu-pgo-20210103T1125.tar.zst".to_string(),
-				sha256: "cff5b10fab51f9e774bb9e0f9878f1ba2703d005618ce0899905d0a1dac31b45".to_string(),
+				url: "https://github.com/indygreg/python-build-standalone/releases/download/20210228/cpython-3.9.2-x86_64-unknown-linux-gnu-pgo-20210228T1503.tar.zst".to_string(),
+				sha256: "a94ee8ccce2e15915bf61077a24454b3a264e0728a74f3fff5d10446cd0d0811".to_string(),
 			},
 			target_triple: "x86_64-unknown-linux-gnu".to_string(),
 			supports_prebuilt_extension_modules: true,
@@ -87,8 +87,8 @@ pub static PYTHON_DISTRIBUTIONS: Lazy<PythonDistributionCollection> = Lazy::new(
 		PythonDistributionRecord {
 			python_major_minor_version: "3.8".to_string(),
 			location: PythonDistributionLocation::Url {
-				url: "https://github.com/indygreg/python-build-standalone/releases/download/20210103/cpython-3.8.7-x86_64-unknown-linux-musl-noopt-20210103T1125.tar.zst".to_string(),
-				sha256: "5b9f4bcfe550b8ecdb0f74f58b08c2b0fcf7f7ab49d9f0f00865691667d47413".to_string(),
+				url: "https://github.com/indygreg/python-build-standalone/releases/download/20210228/cpython-3.8.8-x86_64-unknown-linux-musl-noopt-20210228T1503.tar.zst".to_string(),
+				sha256: "31ca3fad65ba39071712aa6c1788e787bcc0f2140303927c4957e1395d6bf862".to_string(),
 			},
 			target_triple: "x86_64-unknown-linux-musl".to_string(),
 			supports_prebuilt_extension_modules: false,
@@ -96,8 +96,8 @@ pub static PYTHON_DISTRIBUTIONS: Lazy<PythonDistributionCollection> = Lazy::new(
 		PythonDistributionRecord {
 			python_major_minor_version: "3.9".to_string(),
 			location: PythonDistributionLocation::Url {
-				url: "https://github.com/indygreg/python-build-standalone/releases/download/20210103/cpython-3.9.1-x86_64-unknown-linux-musl-noopt-20210103T1125.tar.zst".to_string(),
-				sha256: "fcf77b7cc208d302abb459ba140ba2668c977499b47817a671405df8ccf0846a".to_string(),
+				url: "https://github.com/indygreg/python-build-standalone/releases/download/20210228/cpython-3.9.2-x86_64-unknown-linux-musl-noopt-20210228T1503.tar.zst".to_string(),
+				sha256: "eca125ef4ec230ea3daaed912e506af2bb07510da6309cbbdb3c862677da71e2".to_string(),
 			},
 			target_triple: "x86_64-unknown-linux-musl".to_string(),
 			supports_prebuilt_extension_modules: false,
@@ -114,8 +114,8 @@ pub static PYTHON_DISTRIBUTIONS: Lazy<PythonDistributionCollection> = Lazy::new(
 		PythonDistributionRecord {
 			python_major_minor_version: "3.8".to_string(),
 			location: PythonDistributionLocation::Url {
-				url: "https://github.com/indygreg/python-build-standalone/releases/download/20210103/cpython-3.8.7-i686-pc-windows-msvc-shared-pgo-20210103T1125.tar.zst".to_string(),
-				sha256: "c4d577718d02faf508b53458a7039ce6f947897f8be0b689fb138acbf41b496c".to_string(),
+				url: "https://github.com/indygreg/python-build-standalone/releases/download/20210228/cpython-3.8.8-i686-pc-windows-msvc-shared-pgo-20210228T1503.tar.zst".to_string(),
+				sha256: "e1d4cf5fb72b5c613edac4ddf57eb576b7a4f4f2b8562b7c95e06fe4aad6f54d".to_string(),
 			},
 			target_triple: "i686-pc-windows-msvc".to_string(),
 			supports_prebuilt_extension_modules: true,
@@ -123,8 +123,8 @@ pub static PYTHON_DISTRIBUTIONS: Lazy<PythonDistributionCollection> = Lazy::new(
 		PythonDistributionRecord {
 			python_major_minor_version: "3.9".to_string(),
 			location: PythonDistributionLocation::Url {
-				url: "https://github.com/indygreg/python-build-standalone/releases/download/20210103/cpython-3.9.1-i686-pc-windows-msvc-shared-pgo-20210103T1125.tar.zst".to_string(),
-				sha256: "e31dbf2fee5f0be6992088b925b8e11c651b7afbcf5111cb888d069bb3273575".to_string(),
+				url: "https://github.com/indygreg/python-build-standalone/releases/download/20210228/cpython-3.9.2-i686-pc-windows-msvc-shared-pgo-20210228T1503.tar.zst".to_string(),
+				sha256: "362a658b295c1161ba744d194254064d8eb36a20424918b0adb459cb74bc07f0".to_string(),
 			},
 			target_triple: "i686-pc-windows-msvc".to_string(),
 			supports_prebuilt_extension_modules: true,
@@ -132,8 +132,8 @@ pub static PYTHON_DISTRIBUTIONS: Lazy<PythonDistributionCollection> = Lazy::new(
 		PythonDistributionRecord {
 			python_major_minor_version: "3.8".to_string(),
 			location: PythonDistributionLocation::Url {
-				url: "https://github.com/indygreg/python-build-standalone/releases/download/20210103/cpython-3.8.7-x86_64-pc-windows-msvc-shared-pgo-20210103T1125.tar.zst".to_string(),
-				sha256: "8fbd66f2ff97192f7c48fd3d00de3330f894531ccbe01205d86b6ed5f4108faa".to_string(),
+				url: "https://github.com/indygreg/python-build-standalone/releases/download/20210228/cpython-3.8.8-x86_64-pc-windows-msvc-shared-pgo-20210228T1503.tar.zst".to_string(),
+				sha256: "5d9c719db2162d9ee48b6b724c8ecfc99c6d6e138799f4e7a2fd9bdefd553fda".to_string(),
 			},
 			target_triple: "x86_64-pc-windows-msvc".to_string(),
 			supports_prebuilt_extension_modules: true,
@@ -141,8 +141,8 @@ pub static PYTHON_DISTRIBUTIONS: Lazy<PythonDistributionCollection> = Lazy::new(
 		PythonDistributionRecord {
 			python_major_minor_version: "3.9".to_string(),
 			location: PythonDistributionLocation::Url {
-				url: "https://github.com/indygreg/python-build-standalone/releases/download/20210103/cpython-3.9.1-x86_64-pc-windows-msvc-shared-pgo-20210103T1125.tar.zst".to_string(),
-				sha256: "a98f2c0e03d2aeac71d956e3a83c550552d069dde3a4c3b11308504979c5db35".to_string(),
+				url: "https://github.com/indygreg/python-build-standalone/releases/download/20210228/cpython-3.9.2-x86_64-pc-windows-msvc-shared-pgo-20210228T1503.tar.zst".to_string(),
+				sha256: "c960f615036a0ab0f4b608dcf32ff30c6647bf39843e8033fc4b9a56fc3eed7a".to_string(),
 			},
 			target_triple: "x86_64-pc-windows-msvc".to_string(),
 			supports_prebuilt_extension_modules: true,
@@ -152,8 +152,8 @@ pub static PYTHON_DISTRIBUTIONS: Lazy<PythonDistributionCollection> = Lazy::new(
 		PythonDistributionRecord {
 			python_major_minor_version: "3.8".to_string(),
 			location: PythonDistributionLocation::Url {
-				url: "https://github.com/indygreg/python-build-standalone/releases/download/20210103/cpython-3.8.7-i686-pc-windows-msvc-static-noopt-20210103T1125.tar.zst".to_string(),
-				sha256: "03ff84890db1bbf3f6351639c0c682dcc4f7625f88f5f9588cae6f5429d3382c".to_string(),
+				url: "https://github.com/indygreg/python-build-standalone/releases/download/20210228/cpython-3.8.8-i686-pc-windows-msvc-static-noopt-20210228T1503.tar.zst".to_string(),
+				sha256: "eeb7f15559e86da372ba49101e400c06da13da0366c5d393776042f03bceafa1".to_string(),
 			},
 			target_triple: "i686-pc-windows-msvc".to_string(),
 			supports_prebuilt_extension_modules: false,
@@ -161,8 +161,8 @@ pub static PYTHON_DISTRIBUTIONS: Lazy<PythonDistributionCollection> = Lazy::new(
 		PythonDistributionRecord {
 			python_major_minor_version: "3.9".to_string(),
 			location: PythonDistributionLocation::Url {
-				url: "https://github.com/indygreg/python-build-standalone/releases/download/20210103/cpython-3.9.1-i686-pc-windows-msvc-static-noopt-20210103T1125.tar.zst".to_string(),
-				sha256: "0bd2965f95093a5b89dc44a236d7955ed5d51721658c22cf45580039e0f795af".to_string(),
+				url: "https://github.com/indygreg/python-build-standalone/releases/download/20210228/cpython-3.9.2-i686-pc-windows-msvc-static-noopt-20210228T1503.tar.zst".to_string(),
+				sha256: "a42c198a55963357fdd9c340de7b4e52ed8a9a92ce459b6c795506f639623816".to_string(),
 			},
 			target_triple: "i686-pc-windows-msvc".to_string(),
 			supports_prebuilt_extension_modules: false,
@@ -170,8 +170,8 @@ pub static PYTHON_DISTRIBUTIONS: Lazy<PythonDistributionCollection> = Lazy::new(
 		PythonDistributionRecord {
 			python_major_minor_version: "3.8".to_string(),
 			location: PythonDistributionLocation::Url {
-				url: "https://github.com/indygreg/python-build-standalone/releases/download/20210103/cpython-3.8.7-x86_64-pc-windows-msvc-static-noopt-20210103T1125.tar.zst".to_string(),
-				sha256: "246a071e249a17018c4ace2fe80163fff43891b94193b361f80a0ad69008284e".to_string(),
+				url: "https://github.com/indygreg/python-build-standalone/releases/download/20210228/cpython-3.8.8-x86_64-pc-windows-msvc-static-noopt-20210228T1503.tar.zst".to_string(),
+				sha256: "42378ca1788018cff9ee077412c3b2fa3d05829377e7f3448d4b9f6cba481ee1".to_string(),
 			},
 			target_triple: "x86_64-pc-windows-msvc".to_string(),
 			supports_prebuilt_extension_modules: false,
@@ -179,8 +179,8 @@ pub static PYTHON_DISTRIBUTIONS: Lazy<PythonDistributionCollection> = Lazy::new(
 		PythonDistributionRecord {
 			python_major_minor_version: "3.9".to_string(),
 			location: PythonDistributionLocation::Url {
-				url: "https://github.com/indygreg/python-build-standalone/releases/download/20210103/cpython-3.9.1-x86_64-pc-windows-msvc-static-noopt-20210103T1125.tar.zst".to_string(),
-				sha256: "b7dea448bebf2b73c5da8f6979f3db4ca17ebdfadfd81f1c67027220b3ef68e5".to_string(),
+				url: "https://github.com/indygreg/python-build-standalone/releases/download/20210228/cpython-3.9.2-x86_64-pc-windows-msvc-static-noopt-20210228T1503.tar.zst".to_string(),
+				sha256: "66cb7aef8f07f6a787f54b7a9c733cdfd89ca8fa8f5cfe643fbae9e5713052b1".to_string(),
 			},
 			target_triple: "x86_64-pc-windows-msvc".to_string(),
 			supports_prebuilt_extension_modules: false,
@@ -190,8 +190,8 @@ pub static PYTHON_DISTRIBUTIONS: Lazy<PythonDistributionCollection> = Lazy::new(
 		PythonDistributionRecord {
 			python_major_minor_version: "3.8".to_string(),
 			location: PythonDistributionLocation::Url {
-				url: "https://github.com/indygreg/python-build-standalone/releases/download/20210103/cpython-3.8.7-x86_64-apple-darwin-pgo-20210103T1125.tar.zst".to_string(),
-				sha256: "d76300bb7967b7e5e361a092964d8623141fd5b1e41abae20d7ac6fb87e56c92".to_string(),
+				url: "https://github.com/indygreg/python-build-standalone/releases/download/20210228/cpython-3.8.8-x86_64-apple-darwin-pgo-20210228T1503.tar.zst".to_string(),
+				sha256: "714ad6215900393528a7fcf38a5266006f613fe506f266ec5b7492ae986aba10".to_string(),
 			},
 			target_triple: "x86_64-apple-darwin".to_string(),
 			supports_prebuilt_extension_modules: true,
@@ -199,8 +199,8 @@ pub static PYTHON_DISTRIBUTIONS: Lazy<PythonDistributionCollection> = Lazy::new(
 		PythonDistributionRecord {
 			python_major_minor_version: "3.9".to_string(),
 			location: PythonDistributionLocation::Url {
-				url: "https://github.com/indygreg/python-build-standalone/releases/download/20210103/cpython-3.9.1-x86_64-apple-darwin-pgo-20210103T1125.tar.zst".to_string(),
-				sha256: "3dac2b81542180e119899c94ea55edc77e3a210a87f56686b38718322e8c6fb5".to_string(),
+				url: "https://github.com/indygreg/python-build-standalone/releases/download/20210228/cpython-3.9.2-x86_64-apple-darwin-pgo-20210228T1503.tar.zst".to_string(),
+				sha256: "bcb19ea30244d01b91a5ac5e51ecf15c0c0bd9b1fc86177d2c595657dc396ef1".to_string(),
 			},
 			target_triple: "x86_64-apple-darwin".to_string(),
 			supports_prebuilt_extension_modules: true,