+++ docs/history.rst
@@ -164,8 +164,9 @@ New Features
 Other Relevant Changes
 ^^^^^^^^^^^^^^^^^^^^^^
 
-* Python distributions upgraded to CPython 3.8.7 and 3.9.1 (from 3.8.6 and 3.9.0).
+* Python distributions upgraded to CPython 3.8.8 and 3.9.2 (from 3.8.6 and 3.9.0).
   See https://github.com/indygreg/python-build-standalone/releases/tag/20210103
+  and https://github.com/indygreg/python-build-standalone/releases/tag/20210227
   for a full list of changes in these distributions.
 * CI has been moved from Azure Pipelines to GitHub Actions.
 * Low level code in the ``pyembed`` crate for loading and indexing resources